First pass at hqc-192

+++ crypto_kem/hqc-rmrs-192/clean/code.c
@@ -21,7 +21,7 @@
  * @param[out] em Pointer to an array that is the tensor code word
  * @param[in] m Pointer to an array that is the message
  */
-void PQCLEAN_HQCRMRS192_CLEAN_code_encode(uint8_t *em, const uint8_t *m) {
+void PQCLEAN_HQCRMRS192_CLEAN_code_encode(uint64_t *em, const uint8_t *m) {
     uint8_t tmp[VEC_N1_SIZE_BYTES] = {0};
 
     PQCLEAN_HQCRMRS192_CLEAN_reed_solomon_encode(tmp, m);
@@ -37,7 +37,7 @@ void PQCLEAN_HQCRMRS192_CLEAN_code_encode(uint8_t *em, const uint8_t *m) {
  * @param[out] m Pointer to an array that is the message
  * @param[in] em Pointer to an array that is the code word
  */
-void PQCLEAN_HQCRMRS192_CLEAN_code_decode(uint8_t *m, const uint8_t *em) {
+void PQCLEAN_HQCRMRS192_CLEAN_code_decode(uint8_t *m, const uint64_t *em) {
     uint8_t tmp[VEC_N1_SIZE_BYTES] = {0};
 
     PQCLEAN_HQCRMRS192_CLEAN_reed_muller_decode(tmp, em);
